fix: about menu
